feat: Adiciona validação de credenciais no formulário de login

+++ src/js/render-projetos.js
@@ -1,9 +1,12 @@
+// Função para renderizar os projetos com modo de admin e modo usuário
 export async function renderProjetos(admin = false) {
+  // Container de projetos
   let containerProjetos = document.getElementById("container-projetos");
   let response = await fetch("/api/projetos");
   let data = await response.json();
+  // Limpa o container de projetos
   containerProjetos.innerHTML = "";
-
+  // Para cada projeto, cria um card
   data.forEach((projeto) => {
     let cardProjeto = document.createElement("div");
     cardProjeto.className =
@@ -17,6 +20,7 @@ export async function renderProjetos(admin = false) {
       <a href="${
         projeto.link
       }" class="text-blue-800 underline" target="_blank">Link</a>
+      
       ${
         admin
           ? `<div class="flex gap-2 mt-2">
@@ -26,16 +30,28 @@ export async function renderProjetos(admin = false) {
           : ""
       }
     `;
-
+    // Container armazena os cards
     containerProjetos.appendChild(cardProjeto);
   });
 
+  // Se estiver no modo admin
   if (admin) {
-    // Adiciona funções aos botões de editar
+    // Valida se há projetos cadastrados
+    if (containerProjetos.childElementCount === 0) {
+      containerProjetos.innerHTML = `
+      <p class="text-2xl text-center text-black font-bold animate-pulse">Nenhum projeto cadastrado</p>
+      `;
+    }
+    // Adiciona funções aos botões de editar (em breve...)
     // Adiciona funções aos botões de excluir
     document.querySelectorAll(".btn-excluir").forEach((btn) => {
       btn.addEventListener("click", async () => {
         const id = btn.dataset.id;
+        const confirmar = confirm(
+          "Tem certeza que deseja excluir esse projeto?"
+        );
+        if (!confirmar) return; // se clicar em "Cancelar", nada acontece
+
         await fetch(`/api/projetos/${id}`, { method: "DELETE" });
         renderProjetos(true); // recarrega lista no modo admin
       });
